pyproject.toml
[tool.poetry]
name = "touchstream"
version = "0.1.0"
description = "API Integration"
authors = ["dimple"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.3"
requests = "^2.28.1"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
django-redis = "^5.2.0"
psycopg2 = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
